+++ src/org/zeroxlab/benchmark/CaseDrawText.java
@@ -61,7 +61,7 @@ public class CaseDrawText extends Case{
         }
 
         String result = "";
-        long total = 0;
+        float total = 0;
         int length = mResult.length;
 
         for (int i = 0; i < length; i++) {
